+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/ReachFiveError.kt
@@ -7,6 +7,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class ReachFiveError(
     override val message: String,
+    val code: Int? = null,
     val exception: Exception? = null,
     val data: ReachFiveApiError? = null
 ) : java.lang.Exception(message), Parcelable {